% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal.R
\name{\%==\%}
\alias{\%==\%}
\alias{equals}
\title{Vector equality assuming NA matches a value}
\usage{
a \%==\% b

equals(a, b)
}
\arguments{
\item{a}{Vector to compare}

\item{b}{Vector to compare}
}
\value{
Logical if \code{a} and \code{b} are equal, assuming NA is equal to anything.
}
\description{
Provides an alternate equality assuming NA is equal to anything,
including another NA.
}
\section{Functions}{
\itemize{
\item \code{a \%==\% b}: Operator for equality

}}
\examples{
\dontrun{
c(1,2,3) \%==\% c(1,NA,3)
# [1] TRUE TRUE TRUE
}
}
