% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meld.R
\name{meld}
\alias{meld}
\title{Meld variables in data frame}
\usage{
meld(x, ..., remove = TRUE)
}
\arguments{
\item{x}{A data frame to operate on.}

\item{...}{Named lists (e.g., \verb{new1=c("old1","old2)}). Specifies the new variable name (\code{new1}) that
results from melding the two old variables (\code{old1} and \code{old2}). List of old variables can be
more than 2.}

\item{remove}{Logical, should the original variables be removed from the data frame (default is TRUE).}
}
\value{
A data frame with the old variables melded into new variables and old variables removed (assuming
\code{remove=TRUE}).
}
\description{
Combine variables in a data frame into a single, new variable (coalesce) with warnings
unless the fields are equal (or one is NA).
}
\examples{
meld(data.frame(a=c(1,2,3), b=c(NA,2,3)), new=c("a","b"))
#   new
# 1   1
# 2   2
# 3   3
}
