% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{meld_collisions}
\alias{meld_collisions}
\title{Meld column collisions in a data frame}
\usage{
meld_collisions(x)
}
\arguments{
\item{x}{A data frame with (possibly) column names with collisions.}
}
\value{
A data frame with the repeated columns melded.
}
\description{
Combine columns that appear to be collisions (ending in .x and
.y) by melding them.
}
\details{
When using dplyr, sometimes you do not want to join by a column because
the data may be partially missing. When doing this, both columns are in the result
but appended with \code{.x} and \code{.y}. This function will identify these column pairs
and \code{\link[=meld]{meld()}} them into a single variable (without the \code{.x} or \code{.y} extension).

Note that the side effect of \code{\link[=meld]{meld()}} is to \code{\link[base:stop]{base::stop()}} if any values are not
matched (other than NA) in the two columns. This provides a safe way to combine
data if you are not sure about it's contents.
}
\examples{
\dontrun{
meld_collisions(dplyr::left_join(iris, iris, by="Species"))
}
}
