% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meld.R
\name{\%^\%}
\alias{\%^\%}
\alias{vec_meld}
\title{Meld two vectors, warning if unequal}
\usage{
a \%^\% b

vec_meld(a, b)
}
\arguments{
\item{a}{Vector to meld}

\item{b}{Vector to meld}
}
\value{
A vector melding a and b (see \code{\link[dplyr:coalesce]{dplyr::coalesce()}}).
}
\description{
The operator \code{\link[dplyr:coalesce]{dplyr::coalesce()}} is useful for combining
vectors, however if the values do not agree then coalesce use the first
non-NA result. This function will print a warning if there
are value mismatches.
}
\section{Functions}{
\itemize{
\item \code{a \%^\% b}: Binary operator for two vectors

}}
\examples{
\dontrun{
c("A","B","C") \%^\% c("A",NA,"C")
# [1] "A" "B" "C"
}
}
